requirements.txt
pyautogui==0.9.54
mouse==0.7.1
pynput==1.7.6
pyyaml==6.0.1
